examples/dynamic-lora-sidecar/requirements.txt
aiohttp==3.10.10
pyyaml==6.0.2
requests==2.32.3
watchdog==5.0.3